refactor(terminal): move restore orchestration types
